requirements.txt
pillow==11.3.0
pandas==2.3.2
ultralytics==8.3.202
opencv-python~=4.12.0.88
numpy~=2.2.6
pip~=25.1.1
filelock~=3.19.1
pytz~=2025.2
charset-normalizer~=3.4.3
setuptools~=80.9.0
fsspec~=2025.9.0
matplotlib~=3.10.6
MarkupSafe~=3.0.2
scipy~=1.16.2
python-dateutil~=2.9.0.post0
six~=1.17.0
opencv-python==4.12.0.88